Move media sanitize function to the separate file
